+++ src/TGUI/Widget.cpp
@@ -930,10 +930,11 @@ namespace tgui
 
     /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-    void Widget::updateTime(Duration elapsedTime)
+    bool Widget::updateTime(Duration elapsedTime)
     {
         m_animationTimeElapsed += elapsedTime;
 
+        const bool screenRefreshRequired = !m_showAnimations.empty();
         for (unsigned int i = 0; i < m_showAnimations.size();)
         {
             if (m_showAnimations[i]->update(elapsedTime))
@@ -941,6 +942,8 @@ namespace tgui
             else
                 i++;
         }
+
+        return screenRefreshRequired;
     }
 
     /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////