+++ src/TGUI/Widgets/TextBox.cpp
@@ -1956,9 +1956,9 @@ namespace tgui
 
     /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-    void TextBox::updateTime(Duration elapsedTime)
+    bool TextBox::updateTime(Duration elapsedTime)
     {
-        Widget::updateTime(elapsedTime);
+        bool screenRefreshRequired = Widget::updateTime(elapsedTime);
 
         // Only show/hide the caret every half second
         if (m_animationTimeElapsed >= getEditCursorBlinkRate())
@@ -1971,7 +1971,11 @@ namespace tgui
 
             // Too slow for double clicking
             m_possibleDoubleClick = false;
+
+            screenRefreshRequired = true;
         }
+
+        return screenRefreshRequired;
     }
 
     /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////